+++ src/components/pages/apps/genes/gsea/bubble/gsea-bubble-provider.tsx
@@ -1,9 +1,18 @@
-import { createContext, useContext, useState, type ReactNode } from 'react'
+import {
+  createContext,
+  useContext,
+  useMemo,
+  useState,
+  type ReactNode,
+} from 'react'
 
 import { makeUuid } from '@/lib/id'
 
+import { ColorMap, getColorMap } from '@/lib/color/colormap'
+import { argsort } from '@/lib/math/argsort'
 import { produce } from 'immer'
 import { IBasePlot } from '../../../matcalc/history/history-provider/plot'
+import { useGseaBubbleSettings } from './gsea-bubble-settings-store'
 import {
   DEFAULT_GSEA_DOT_PROPS,
   type IGseaBubbleDisplayOptions,
@@ -18,13 +27,24 @@ export interface IGseaBubble {
 
 export interface IGseaBubblePlot extends IBasePlot {
   style: 'gsea-bubble-plot'
-  gseaDot: IGseaBubble
+  gseaBubble: IGseaBubble
   props: IGseaBubbleDisplayOptions
+}
+
+export interface IBubblePoint {
+  x: number
+  y: number
+  p: number
+  color: string
+  size: number
+  r: number
+  label: string
 }
 
 export interface GseaBubblePropsContextType {
   displayProps: IGseaBubbleDisplayOptions
   plot: IGseaBubblePlot
+  points: IBubblePoint[]
   setPlot: (plot: IGseaBubblePlot) => void
 }
 
@@ -76,13 +96,25 @@ export function newGseaBubblePlot(
     id: makeUuid(),
     style,
     name,
-    gseaDot,
+    gseaBubble: gseaDot,
     groupRows: [],
     props,
     actions,
     type: 'plot',
     createdAt: new Date().toISOString(),
   }
+}
+
+function getColor(
+  log10pvalue: number,
+  maxlog10pvalue: number,
+  colorMap: ColorMap
+) {
+  const f = Math.min(log10pvalue / maxlog10pvalue, 1)
+
+  const color = colorMap.getHexColor(f)
+
+  return color
 }
 
 export function GseaBubbleProvider({
@@ -94,11 +126,80 @@ export function GseaBubbleProvider({
   children: ReactNode
 }) {
   const [_plot, setPlot] = useState<IGseaBubblePlot | undefined>(plot)
+  const { settings } = useGseaBubbleSettings()
+
+  const points = useMemo(() => {
+    if (!_plot) {
+      return []
+    }
+
+    let ids = _plot!.gseaBubble.ids
+    let nes = _plot!.gseaBubble.nes.values
+    let sizes = _plot!.gseaBubble.sizes.values
+    let log10pvalues = _plot!.gseaBubble.log10pvalues.values
+
+    let idx: number[] = []
+
+    switch (settings.sortBy) {
+      case 'nes':
+        idx = argsort(nes, true)
+
+        break
+      case 'size':
+        idx = argsort(sizes, true)
+        break
+      case 'pvalue':
+        idx = argsort(log10pvalues, true)
+        break
+
+      default:
+        break
+    }
+
+    if (idx.length > 0) {
+      nes = idx.map((i) => nes[i])
+      sizes = idx.map((i) => sizes[i])
+      log10pvalues = idx.map((i) => log10pvalues[i])
+      ids = idx.map((i) => ids[i])
+    }
+
+    return nes.map((x, i) => {
+      const size = sizes[i]!
+      const sizeF = Math.min(size / settings.size.maxSize, 1)
+      return {
+        x,
+        y: i + 1,
+        p: log10pvalues[i]!,
+        color: getColor(
+          log10pvalues[i]!,
+          settings.p.range[1],
+          getColorMap(settings.p.cmap)
+        ),
+        size: sizeF,
+        r: sizeF * settings.bubbles.size,
+        label: ids[i]!,
+      }
+    })
+  }, [
+    _plot?.gseaBubble.nes.values,
+    _plot?.gseaBubble.sizes.values,
+    settings.size.maxSize,
+    settings.bubbles.size,
+    settings.p,
+    settings.margin,
+    settings.padding,
+    settings.legend,
+    settings.border,
+    settings.axes,
+    settings.bubbles,
+    settings.sortBy,
+  ])
 
   return (
     <GseaBubbleContext.Provider
       value={{
         plot: _plot,
+        points,
         displayProps: _plot?.props ?? DEFAULT_GSEA_DOT_PROPS,
         setPlot,
       }}